✨ Android Camera2 API 수동 초점 진행중..

+++ app/src/main/java/com/example/finiview/camera/common/event/OnClickCameraExposureEvent.kt
@@ -1,3 +1,3 @@
 package com.example.finiview.camera.common.event
 
-class OnClickCameraExposureEvent(val exposure: Int) : EventBusInterface
+class OnClickCameraExposureEvent(val exposure: Long) : EventBusInterface
